backend/requirements.txt
attrs==19.3.0             # via pytest
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
django==2.2.6
idna==2.8                 # via requests
importlib-metadata==1.5.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.1           # via pytest
pillow
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.6          # via packaging
pytest-django==3.8.0
pytest-pythonpath==0.7.3
pytest==5.3.5             # via pytest-django
requests==2.22.0
six==1.14.0               # via packaging
sorl-thumbnail==12.6.3
urllib3==1.25.6           # via requests
wcwidth==0.1.8            # via pytest
zipp==2.2.0               # via importlib-metadata
mixer==7.1.2
django-debug-toolbar
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.2
PyJWT==2.1.0
django-filter
python-dotenv
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
djoser
# для postgresql
django-environ
flake8
pep8-naming
flake8-broken-line
flake8-return
flake8-isort